Restructure command handling
Reordered checks and moved argument parsing to the end, so that less
errors are falsely reported. This change does not present any fix for
the mismatched quotes problem, just makes it occur less often.

Added comments and refactored the code to make it more readable.

Edited some responses to make them more descriptive or succinct as
needed, or to fix nickb's spelling mistakes.