Decrease chance of fatal failures
Prevent errors that occur asynchronously and can crash the bot.
Remove chained throw when command errors occur. This is needed
because the error manager has been removed.

Also fixes a ton of style things, access modifiers, etc.
Another fix is changing == to += in PointsManager. This was messed up
during the bot rewrite.